I've been working on my custom computer for a while and unfortunately didn't think to start tracking changes to the CPU's design document until now.  Suffice it to say, I began my design thinking about all the various features that might be useful in a computer system and I added them to my CPU.
I will attempt to describe below the up-front design decisions that I
made when I wrote the initial specification and the preliminary
simulator.

I wanted many general purpose registers so that I could make
assembly-level programming easy, and so that it would make optimizations
with the compiler easy (the register allocation scheme should be easier
to implement if there are a lot of registers to choose from to begin
with). So, I ended up creating a machine that will have 32 general
purpose registers. Each register is encoded as a 5-bit value.

I decided that I wanted to potentially have support for up to 64
instructions (as that seems like plenty for a relatively simple design
like mine), which means that opcodes will take up 6 bits.

I wanted to make decoding of instructions easy, so each instruction has
a fixed width and relatively fixed positions for the various fields.
This should make a physical implementation consume less hardware and
therefore run faster.

I wanted my computer to be compatible with existing RAM technologies so
that when I make a physical implementation, I can use normal RAM chips
instead of having to design my own memory. So, I decided to make
instructions an integer multiple of a standard 8-bit byte.

Since most instructions are 3-address instructions, they need at least
15 bits to specify the two operand and one result register, plus another
6 bits (for a total of 21) for the opcode. I also wanted my instruction
size to be a power of 2 (for aesthetic reasons) so my instruction length
is 32-bits. Many instructions therefore will have many unused bits (at
least for the time-being).

I have support for the standard integer arithmetic and logic
instructions (AND, OR, NOT, XOR, etc). And because my instruction size
is already 32-bits, I decided to make my word size the same. So all
general purpose registers on this machine are also 32-bits long. To make
implementation easier, memory is word-addressable instead of
byte-addressable, so that way there is no need to worry about unaligned
accesses as the interface to memory will always grab 32-bits at a time
(although this may have to change in the future if implementing this
feature proves difficult in hardware).

There may be floating point support in the future, but for now, the
intent is to get a basic computer system that is capable of running an
operating system complete with user programs up and running.

There may also be support added later for special instructions to help
with certain tasks (such as atomic compare-and-swaps to help implement
locks), but I will make that decision further down the road.

This ISA was inspired partly by Yale Patt's LC-3 with some cues taken
from the MIPS ISA. My goal with this machine is to create a design that
is as straightforward/easy to understand as possible while still having
some "processing oomph". If I decide to prioritize performance over
understandability in the future, I will create a mostly compatible ISA
and let this ISA stay as easy to understand as possible.